refactor(mon-pix): move QCU proposals
