+++ mon-pix/app/components/proposals/qcu-proposals.js
